refac(c/zk): move `buffer_reader` to `c/zk/plonk/halo2/`
I moved the `BufferReader` to an appropriate location
because the upcoming implementation of `ReadHalo2Params` will use it.
