+++ tachyon/c/zk/plonk/keys/BUILD.bazel
@@ -47,28 +47,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "buffer_reader",
-    hdrs = ["buffer_reader.h"],
-    deps = [
-        "//tachyon/base:logging",
-        "//tachyon/base/buffer:endian_auto_reset",
-        "//tachyon/base/buffer:read_only_buffer",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/math/elliptic_curves:points",
-        "//tachyon/math/finite_fields:prime_field_base",
-        "//tachyon/math/polynomials/univariate:univariate_evaluations",
-        "//tachyon/math/polynomials/univariate:univariate_polynomial",
-        "//tachyon/zk/expressions:expression_factory",
-        "//tachyon/zk/lookup:lookup_argument",
-        "//tachyon/zk/plonk/base:column_key",
-        "//tachyon/zk/plonk/base:phase",
-        "//tachyon/zk/plonk/constraint_system:gate",
-        "//tachyon/zk/plonk/permutation:permutation_argument",
-        "//tachyon/zk/plonk/permutation:permutation_proving_key",
-    ],
-)
-
-tachyon_cc_library(
     name = "endian_auto_reset",
     hdrs = ["endian_auto_reset.h"],
     deps = ["//tachyon/base/buffer"],
@@ -78,11 +56,11 @@ tachyon_cc_library(
     name = "proving_key_impl_base",
     hdrs = ["proving_key_impl_base.h"],
     deps = [
-        ":buffer_reader",
         "//tachyon/base:environment",
         "//tachyon/base:logging",
         "//tachyon/base/buffer",
         "//tachyon/base/files:file_util",
+        "//tachyon/c/zk/plonk/halo2:buffer_reader",
         "//tachyon/zk/plonk/halo2:pinned_verifying_key",
         "//tachyon/zk/plonk/keys:proving_key",
         "@com_google_absl//absl/types:span",
